apps/backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
motor==3.6.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.12
orjson==3.10.10
email-validator==2.1.0.post1
